requirements.txt
Flask==2.3.3
supabase==1.0.4
python-dotenv==1.0.0
boto3==1.28.57
requests==2.31.0
authlib==1.2.1
requests-oauthlib==1.3.1